requirements.txt
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.2.1
cffi==1.14.5
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0

gcloud==0.17.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0



google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.37.0
gunicorn==20.0.4
httplib2==0.18.1
jws==0.1.3
msgpack==1.0.2
oauth2client==3.0.0
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pymongo==3.11.3
pyparsing==2.4.7
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
whitenoise==5.2.0